requirements.txt
numpy==2.2.3
opencv_contrib_python==4.9.0.80
opencv_python==4.11.0.86
Pillow==11.1.0
tensorflow==2.18.0
tensorflow_hub==0.16.1
tensorflow_intel==2.18.0
